DnD: Revamped code to simplify / untangle of internal data handling: - C-ifying and renaming classes DnDURIList / DnDURIObject -> DnDTransferList / DnDTransferObject - Added testcases for DnDTransferList / DnDTransferObject + DnDPath API - Reduced memory footprint - Greatly simplified / stripped down internal data flow of Main side - More (optional) release logging for further diagnosis
Work in progress.

git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@85371 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Additions/WINNT/VBoxTray/VBoxDnD.h
@@ -70,7 +70,7 @@ public:
 
     int Abort(void);
     void SetStatus(Status status);
-    int Signal(const RTCString &strFormat, const void *pvData, uint32_t cbData);
+    int Signal(const RTCString &strFormat, const void *pvData, size_t cbData);
 
 protected:
 
@@ -87,7 +87,7 @@ protected:
     RTSEMEVENT  mEventDropped;
     RTCString   mstrFormat;
     void       *mpvData;
-    uint32_t    mcbData;
+    size_t      mcbData;
 };
 
 class VBoxDnDDropSource : public IDropSource
